+++ movingspeed.cpp
@@ -1,15 +1,15 @@
 #include "movingspeed.h"
 
 
-float MovingSpeed::x_speed = 3;
-float MovingSpeed::y_speed = 3;
+long double MovingSpeed::x_speed = 3;
+long double MovingSpeed::y_speed = 3;
 
 int MovingSpeed::current_power = 0;
-float MovingSpeed::current_speed = 0;
+double MovingSpeed::current_speed = 0;
 
 const int MovingSpeed::division_power_factor_index = 20;
 const int MovingSpeed::max_power = 200;
-const float MovingSpeed::division_factor_speed = 40;
+const double MovingSpeed::division_factor_speed = 40;
 const int MovingSpeed::max_speed[11] = {
     0,
     60,
